Add new class for message parsing used by GCM and FCM receivers. Update packages.
