+++ pipitit-android/src/main/java/com/hyperether/pipitit/push/PipititPushListener.java
@@ -1,4 +1,4 @@
-package com.hyperether.pipitit.firebase;
+package com.hyperether.pipitit.push;
 
 import android.content.Intent;
